# documentation/openapi.yaml
openapi: 3.1.0
info:
  title: foodOrdering
  version: '1.0'
  summary: FoodOrdering API
  description: ''
servers:
  - url: 'http://localhost:3000'
paths:
  /user:
    post:
      summary: Create New User
      operationId: post-user
      responses:
        '200':
          description: User Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userCreateResponse'
              examples:
                New User Bob Fellow:
                  value:
                    id: 12
                    firstName: Bob
                    lastName: Fellow
                    email: bob.fellow@gmail.com
                    dateOfBirth: '1996-08-24'
                    emailVerified: false
                    createDate: '2020-11-18'
            application/xml:
              schema:
                type: object
                properties: {}
        '400':
          description: Missing Required Information
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userCreate'
            examples:
              Create User Bob Fellow:
                value:
                  firstName: Bob
                  lastName: Fellow
                  email: bob.fellow@gmail.com
                  dateOfBirth: '1996-08-24'
          application/xml:
            schema:
              type: object
              properties: {}
        description: Post the necessary fields for the API to create a new user.
      description: Create a new user.
    get:
      summary: ''
      operationId: get-user
    put:
      security:
        bearerAuth: []
      summary: ''
      operationId: put-user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: {}
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userPut'
    delete:
      summary: ''
      operationId: delete-user
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
      parameters: []
      security:
        - bearer: []
...
components:
  schemas:
...
    userPut:
      title: userPut
      type: object
      properties:
        password:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        address:
          type: string
        phone:
          type: string
...
  securitySchemes:
    bearer:
      name: bearer
      type: apiKey
      in: header